backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==10.1.0
qrcode[pil]==7.4.2
reportlab==4.0.7
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
